requirements.txt
absl-py==2.1.0
addict==2.4.0
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-kms==2.16.2
attrs==23.2.0
blobfile==2.1.1
cachetools==5.3.3
ccimport==0.4.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
cffi==1.16.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
chumpy==0.70
click==8.1.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
ConfigArgParse==1.7
coverage==7.4.3
crcmod==1.7
cryptography==42.0.5
cumm-cu113==0.4.11
cycler==0.12.1
Cython==3.0.9
einops==0.6.0
exceptiongroup==1.2.0
filelock==3.12.0
fire==0.6.0
flake8==7.0.0
google-auth==2.28.2
google-auth-oauthlib==0.4.6
grpcio==1.62.1
idna @ file:///croot/idna_1666125576474/work
imageio==2.9.0
imageio-ffmpeg==0.4.3
importlib_metadata==7.0.2
iniconfig==2.0.0
interrogate==1.5.0
iopath==0.1.9
isort==4.3.21
jmespath==0.10.0
json-tricks==3.17.3
kiwisolver==1.4.5
lark==1.1.9
lpips==0.1.4
lxml==4.9.4
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.4.2
mccabe==0.7.0
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
model-index==0.1.11
mrcfile==1.5.0
munch==3.0.0
munkres==1.1.4
networkx==3.1
ninja==1.10.2
numpy==1.22.4
oauthlib==3.2.2
opencv-python==4.7.0.72
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.36
ordered-set==4.1.0
oss2==2.17.0
packaging==24.0
pandas==2.0.3
parameterized==0.9.0
pccm==0.4.11
Pillow==9.4.0
platformdirs==4.2.0
pluggy==1.4.0
plyfile==1.0.3
portalocker==2.8.2
protobuf==4.25.3
psutil==5.9.4
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pycocotools==2.0.7
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
pyflakes==3.2.0
Pygments==2.17.2
PyMCubes==0.1.4
pyparsing==3.1.2
pytest==8.1.1
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytorch-msssim==0.2.1
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.28.2
requests-oauthlib==1.4.0
rich==13.4.2
rsa==4.9
scikit-image==0.19.3
scipy==1.9.1
shapely==2.0.3
six==1.16.0
smplx==0.1.28
spconv-cu113==2.3.3
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==2.4.0
terminaltables==3.1.10
tifffile==2023.7.10
toml==0.10.2
tomli==2.0.1
tqdm==4.65.2
trimesh==3.21.7
typing_extensions @ file:///croot/typing_extensions_1705599297034/work
tzdata==2024.1
urllib3==1.26.18
Werkzeug==3.0.1
xdoctest==1.1.3
xtcocotools==1.14.3
yapf==0.40.2
zipp==3.18.0
